amélioration mapcat/mapcat.inc.php et gestion de l'impact

+++ bo/index.php
@@ -9,13 +9,6 @@ require_once __DIR__.'/user.php';
 
 use Symfony\Component\Yaml\Yaml;
 
-function ganHarvestDate(): ?string { // lit la date de dernière moisson des GAN dans ../dashboard/gans.yaml
-  if (!is_file(__DIR__.'/../dashboard/gans.yaml'))
-    return null;
-  $gans = Yaml::parseFile(__DIR__.'/../dashboard/gans.yaml');
-  return $gans['valid'];
-}
-
 $HTML_HEAD = "<!DOCTYPE html>\n<html><head><title>shomgt-bo@$_SERVER[HTTP_HOST]</title></head><body>\n";
 $HTML_TITLE = "<h2>Interface Back Office (BO) de ShomGT</h2>\n";
 
@@ -24,35 +17,29 @@ $HTML_TITLE = "<h2>Interface Back Office (BO) de ShomGT</h2>\n";
 // Si loggé retourne le login, sinon propose de se loguer et si ! ok alors arrête l'exécuction et propose de s'enregistrer
 $login = Login::login($HTML_HEAD.$HTML_TITLE, 'user.php?action=register');
 
-switch ($_GET['action'] ?? null) {
-  case 'deleteTempOutputBatch': { // revient de runbatch.php et efface, s'il existe, le fichier temporaire créé 
-    $filename = basename($_GET['filename']);
-    //echo "filename=$filename<br>\n";
-    if (is_file(__DIR__."/temp/$filename"))
-      unlink(__DIR__."/temp/$filename");
-    // Explicitement je continue sur le cas suivant
-  }
-  case null: { // Menu après login
+switch ($action = ($_GET['action'] ?? null)) {
+  case null:
+  case 'deleteTempOutputBatch': { // Menu après login avec éventuelle action préalable
+    if ($action == 'deleteTempOutputBatch'){ // revient de runbatch.php et efface, s'il existe, le fichier temporaire créé 
+      $filename = basename($_GET['filename']);
+      //echo "filename=$filename<br>\n";
+      if (is_file(__DIR__."/temp/$filename"))
+        unlink(__DIR__."/temp/$filename");
+    }
     $role = userRole($login);
     $roleDisplay = ($role <> 'normal') ? "/role=$role" : '';
-    $ganHarvestInfo = ($harvestDate = ganHarvestDate()) ? "précédemment moissonné le $harvestDate" : '';
-    $wfsUpdateDate = date('Y-m-d', filemtime(__DIR__."/../shomft/gt.json"));
     echo "$HTML_HEAD<h2>Interface Back Office (BO) de ShomGT ($login$roleDisplay)</h2>\n";
     echo "<ul>\n";
     echo "<li><a href='?action=logout'>Se déloguer</a>, <a href='user.php'>gérer son compte</a></li>\n";
     if ($role == 'restricted') die();
-    echo "<li><a href='runbatch2.php?batch=harvestGan'>Moissonnage du GAN $ganHarvestInfo</a></li>\n";
-    echo "<li><a href='../shomft/updatecolls.php?collections=gt,aem,delmar'>",
-         "Mise à jour de la liste des cartes à partir du serveur WFS du Shom ",
-         "précédemment mise à jour le $wfsUpdateDate</a></li>\n";
     echo "<li><a href='../dashboard/' target='_blank'>",
-         "Identifier les cartes à ajouter/supprimer/actualiser dans le portefeuille ShomGT</a></li>\n";
+         "Identifier les cartes à actualiser grâce au tabeau de bord de l'actualité des cartes</a></li>\n";
     echo "<li><a href='https://diffusion.shom.fr' target='_blank'>",
          "Télécharger une nouvelle version de cartes sur le site du Shom</a></li>\n";
     echo "<li><a href='addmaps.php'>Déposer cette nouvelle version de carte dans le portefeuille</a></li>\n";
     //echo "<li><a href='?action=mapcat'>Modifier le catalogue des cartes</li>\n";
     //echo "<li><a href='?action=obsoleteMap'>Déclarer une carte obsolète</a></li>\n";
-    echo "<li><a href='runbatch2.php?batch=sgupdt'>",
+    echo "<li><a href='runbatch.php?batch=sgupdt'>",
          "Prendre en compte les nouvelles versions dans la consultation des cartes</a></li>\n";
     echo "</ul>\n";
     if ($role <> 'admin') die();
